# docs/openapi/api-gateway.json
{
  "components": {
    "schemas": {
...
      "ManualCompleteRequest": {
        "properties": {
          "quantity": {
            "type": "number"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "quantity",
          "reason"
        ],
        "type": "object"
      },
      "ScanRequest": {
        "properties": {
          "barcode": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          }
        },
        "required": [
          "barcode",
          "quantity"
        ],
        "type": "object"
      },
...
      "WorkerTask": {
        "properties": {
          "dokument": {
            "type": "string"
          },
          "due_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "lokacija": {
            "type": "string"
          },
          "progress": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "stavke_total": {
            "type": "integer"
          }
        },
        "type": "object"
      },
      "WorkerTaskDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/WorkerTask"
          },
          {
            "properties": {
              "stavke": {
                "items": {
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "naziv": {
                      "type": "string"
                    },
                    "obradjena_kolicina": {
                      "type": "number"
                    },
                    "status": {
                      "type": "string"
                    },
                    "trazena_kolicina": {
                      "type": "number"
                    }
                  },
                  "type": "object"
                },
                "type": "array"
              }
            },
            "type": "object"
          }
        ]
      },
      "ZaduznicaAssignment": {
        "properties": {
          "due_at": {
            "format": "date-time",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ZaduznicaItemPayload"
            },
            "type": "array"
          },
          "magacioner_id": {
            "format": "uuid",
            "type": "string"
          },
          "priority": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ZaduznicaCreateRequest": {
        "properties": {
          "assignments": {
            "items": {
              "$ref": "#/components/schemas/ZaduznicaAssignment"
            },
            "type": "array"
          },
          "trebovanje_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
      "ZaduznicaItemPayload": {
        "properties": {
          "quantity": {
            "type": "number"
          },
          "trebovanje_stavka_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "trebovanje_stavka_id",
          "quantity"
        ],
        "type": "object"
      },
      "ZaduznicaReassignRequest": {
        "properties": {
          "target_magacioner_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "target_magacioner_id"
        ],
        "type": "object"
      },
      "ZaduznicaStatusUpdate": {
        "properties": {
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Magacin API Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/catalog/articles": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 25,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "items": {
                      "items": {
                        "properties": {
                          "aktivan": {
                            "type": "boolean"
                          },
                          "barkodovi": {
                            "items": {
                              "properties": {
                                "is_primary": {
                                  "type": "boolean"
                                },
                                "value": {
                                  "type": "string"
                                }
                              },
                              "required": [
                                "value"
                              ],
                              "type": "object"
                            },
                            "type": "array"
                          },
                          "id": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "jedinica_mjere": {
                            "type": "string"
                          },
                          "naziv": {
                            "type": "string"
                          },
                          "sifra": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "sifra",
                          "naziv",
                          "jedinica_mjere",
                          "aktivan",
                          "barkodovi"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "page_size": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "items",
                    "total",
                    "page",
                    "page_size"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Catalog article list"
          }
        },
        "summary": "List catalog articles",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/catalog/articles/{article_id}": {
      "patch": {
        "parameters": [
          {
            "in": "path",
            "name": "article_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "aktivan": {
                    "type": "boolean"
                  },
                  "barkodi": {
                    "items": {
                      "properties": {
                        "is_primary": {
                          "type": "boolean"
                        },
                        "value": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "value"
                      ],
                      "type": "object"
                    },
                    "type": "array"
                  },
                  "jedinica_mjere": {
                    "type": "string"
                  },
                  "naziv": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "aktivan": {
                      "type": "boolean"
                    },
                    "barkodovi": {
                      "items": {
                        "properties": {
                          "is_primary": {
                            "type": "boolean"
                          },
                          "value": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "value"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "jedinica_mjere": {
                      "type": "string"
                    },
                    "naziv": {
                      "type": "string"
                    },
                    "sifra": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "sifra",
                    "naziv",
                    "jedinica_mjere",
                    "aktivan",
                    "barkodovi"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Updated article"
          }
        },
        "summary": "Update catalog article",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/worker/tasks/{zaduznica_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "zaduznica_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerTaskDetail"
                }
              }
            },
            "description": "Task detail"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Worker task detail"
      }
    },
    "/api/worker/tasks/{zaduznica_stavka_id}/complete-manual": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "zaduznica_stavka_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted"
          }
        },
        "summary": "Manual completion"
      }
    },
    "/api/worker/tasks/{zaduznica_stavka_id}/scan": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "zaduznica_stavka_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted"
          }
        },
        "summary": "Register scan"
      }
    },
    "/api/zaduznice": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZaduznicaCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "summary": "Create zaduznice"
      }
    },
    "/api/zaduznice/{zaduznica_id}/reassign": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "zaduznica_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZaduznicaReassignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Updated"
          }
        },
        "summary": "Reassign zaduznica"
      }
    },
    "/api/zaduznice/{zaduznica_id}/status": {
      "patch": {
        "parameters": [
          {
            "in": "path",
            "name": "zaduznica_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZaduznicaStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Updated"
          }
        },
        "summary": "Update zaduznica status"
      }
    }
  }
}